fix: rename arrows for page left and right
